+++ src/frontend/apps/impress/src/features/docs/doc-editor/stores/index.ts
@@ -1,3 +1,2 @@
-export * from './useDocStore';
 export * from './useHeadingStore';
 export * from './usePanelEditorStore';
