+++ src/frontend/apps/impress/src/features/docs/doc-management/index.ts
@@ -1,5 +1,6 @@
 export * from './api';
 export * from './components';
 export * from './hooks';
+export * from './stores';
 export * from './types';
 export * from './utils';
