+++ src/frontend/apps/impress/src/features/docs/doc-management/stores/index.tsx
@@ -1,0 +1,1 @@
+export * from './useDocStore';
